[WIP] Add Compatibility rendering method fallbacks and warnings to 3D demos
This adjusts visuals automatically depending on the rendering
method in use and adds warnings for unsupported features.

Since web exports always use the Compatibility rendering method,
this leads to a better experience out of the box when exporting
demos to the web.

This commit depends on the engine PR adding
`Engine.get_current_rendering_method()` (it will not work on
4.2.x or `master`).